Update deprecation warning about writer_opts.
Previously it only specifiec that writer_opts should be passed into the
renderer_context instead of supplied as an argument to render. This is
because, under normal circumstances, users won't be calling `render`
directly. In the simple cases where you know you want to render the data
as CSV, renderer_context may be the right place to add these options.
However, if you want to provide renderings in several formats, not all
renderers should receive writer_opts in renderer_context. In this case
it would be more appropriate to create a custom CSVRenderer and set the
writer_opts attribute.