Allow Generator data in CSVStreamingRenderer
CSVStreamingRenderer does not support a data source of a Generator object, but the Renderer works properly with it as tabelize works as a generator along with CSVStreamingRenderer itself.  This makes it so that if the data passed in is a Generator object, it does not need to go through the conversion to a list.